ACPI/IORT: Support paravirtualized IOMMU
To describe the virtual topology in relation to a virtio-iommu device,
ACPI-based systems use a "paravirtualized IOMMU" IORT node. Add support
for it. The IORT specification reserves the node type 0x80, but doesn't
actually describe it. The layout is described by the virtio-iommu
specification.

For guests that don't support device tree, the host must pass an ACPI
table to the guest, that describes the relation between virtio-iommu and
endpoints in the virtual machine. The table maps PCI RIDs and integrated
devices to IOMMU device IDs, telling the IOMMU driver which endpoints it
manages.

We choose IORT for paravirtualized IOMMU description among the three
existing ACPI tables for I/O Remapping (Intel DMAR, AMD IVRS and Arm
IORT), because it is the easiest to extend and to implement in the device
driver.

Given that the new node type is 0x80 (as an IORT vendor-specific ID), it
doesn't fit in the NODE_TYPE mask anymore and requires small change in the
way node types are filtered.

Signed-off-by: Jean-Philippe Brucker <jean-philippe.brucker@arm.com>