arch/x86: Only try EFI reboot if booted via EFI
As well as checking for the presence of "noefi" field via
efi_runtime_disabled() also check that the kernel was booted with EFI.

This fixes an issue on Cloud Hypervisor (which uses ACPI
Hardware-Reduced) where the kernel tries to reboot via EFI even though
the kernel was loaded directly into the VM via the PVH entry point. It
should try and use ACPI instead.

Signed-off-by: Rob Bradford <robert.bradford@intel.com>