virtio-net: Add feature check for VIRTIO_NET_CTRL_GUEST_OFFLOADS
Ensure the guest advertises support for VIRTIO_NET_CTRL_GUEST_OFFLOADS
via VIRTIO_NET_F_CTRL_GUEST_OFFLOADS feature before attempting to issues
the request. This fixes an issue on the Cloud Hypervisor VMM which does
not support this functionality or advertise this feature however the
ethtool logic in systemd-udevd was able to trigger an attempt to
change the offloads.

Call Trace:
 ? __kmalloc_node_track_caller+0x20a/0x320
 ? sg_init_table+0x15/0x40
 virtnet_set_guest_offloads+0x57/0x80 [virtio_net]
 virtnet_set_features+0x53/0x70 [virtio_net]
 __netdev_update_features+0x279/0x9f0
 ? vunmap_page_range+0x2c8/0x390
 __netdev_update_features+0x308/0x9f0
 dev_ethtool+0x15b9/0x2950

Signed-off-by: Rob Bradford <robert.bradford@intel.com>